--- CreeDictionary/constants/__init__.py
@@ -1,1 +1,0 @@
-from constants.enums import ParadigmSize, PartOfSpeech, SimpleLexicalCategory
