+++ CreeDictionary/tests/API_tests/model_test.py
@@ -4,8 +4,8 @@ from collections import Iterable
 import pytest
 from hypothesis import assume, given
 
-from API.models import Wordform, filter_cw_wordforms
-from constants.enums import Language
+from API.models import Wordform
+from utils.enums import Language
 from CreeDictionary import settings
 from tests.conftest import random_lemmas
