+++ CreeDictionary/tests/utils_tests/test_layout_combiner.py
@@ -1,5 +1,5 @@
-from constants import ParadigmSize
-from constants.enums import SimpleLC
+from utils import ParadigmSize
+from utils.enums import SimpleLC
 from shared import shared_res_dir
 from utils.paradigm_layout_combiner import import_layouts
