+++ CreeDictionary/utils/__init__.py
@@ -1,3 +1,8 @@
 from .fst_analysis_parser import extract_simple_lc, extract_lemma_and_category
 from .shared_res_dir import shared_res_dir
 from .cree_lev_dist import get_modified_distance
+
+from .enums import *
+from .types import *
+from .vars import *
+from .data_classes import *
