+++ CreeDictionary/utils/enums.py
