+++ CreeDictionary/utils/paradigm_layout_combiner.py
@@ -20,8 +20,8 @@ from typing import Dict, FrozenSet, List, Tuple
 
 import hfstol
 
-from constants import ParadigmSize
-from constants.enums import SimpleLC
+from utils import ParadigmSize
+from utils.enums import SimpleLC
 
 # A raw paradigm layout from Neahttadigisánit.
 Table = List[List[str]]
